rename dense_matrix2d
