+++ source/consumers/dataspaces/DenseMatrix2D.h
